Support masking of atoms and pairs
For practical use in MD or training loops, we need
to be able to deal with atomic positions and pairs
that have been added to pad to a fixed shape. This
commit adds functionality to support this, trying
to keep most of the complexity out of the low-level
parts of the code. Masks are passed to the high-
level calculators, which do most of the work. At
the lower levels, we just have to make sure to
mask out zero distances, which we use as a kind
of "sentinel" value to indicate something that
should be ignored.